refactor(shima2009): add booleans for gen_config, gen_gbxs and gen_supers

+++ examples/boxmodelcollisions/shima2009.py
@@ -143,6 +143,9 @@ def inputfiles(
     config_filename,
     config_params,
     kernel,
+    gen_config,
+    gen_gbxs,
+    gen_supers,
     isfigures,
 ):
     from cleopy import editconfigfile
@@ -158,16 +161,19 @@ def inputfiles(
         savefigpath.mkdir(exist_ok=True)
 
     ### --- copy src_config_filename into tmp and edit parameters --- ###
-    config_filename.unlink(missing_ok=True)  # delete any existing config
-    shutil.copy(src_config_filename, config_filename)
-    editconfigfile.edit_config_params(config_filename, config_params)
+    if gen_config:
+        config_filename.unlink(missing_ok=True)  # delete any existing config
+        shutil.copy(src_config_filename, config_filename)
+        editconfigfile.edit_config_params(config_filename, config_params)
 
     ### --- delete any existing initial conditions --- ###
     yaml = YAML()
     with open(config_filename, "r") as file:
         config = yaml.load(file)
-    Path(config["inputfiles"]["grid_filename"]).unlink(missing_ok=True)
-    Path(config["initsupers"]["initsupers_filename"]).unlink(missing_ok=True)
+    if gen_gbxs:
+        Path(config["inputfiles"]["grid_filename"]).unlink(missing_ok=True)
+    if gen_supers:
+        Path(config["initsupers"]["initsupers_filename"]).unlink(missing_ok=True)
 
     ### --- input binary files generation --- ###
     # equivalent to ``import shima2009_inputfiles`` followed by
@@ -184,6 +190,10 @@ def inputfiles(
         config_filename,
         kernel,
     ]
+    if gen_gbxs:
+        cmd.append("--gen_gbxs")
+    if gen_supers:
+        cmd.append("--gen_supers")
     if isfigures[0]:
         cmd.append("--show_figures")
     if isfigures[1]:
@@ -240,6 +250,9 @@ def plot_results(path2CLEO, config_filename, savefigpath, kernel):
 ### --------------------- RUN EXAMPLE FOR EACH KERNEL ---------------------- ###
 for kernel, [config_filename, config_params] in kernel_configs.items():
     if args.do_inputfiles:
+        gen_config = True
+        gen_gbxs = True
+        gen_supers = True
         inputfiles(
             path2CLEO,
             path2build,
@@ -251,6 +264,9 @@ for kernel, [config_filename, config_params] in kernel_configs.items():
             config_filename,
             config_params,
             kernel,
+            gen_config,
+            gen_gbxs,
+            gen_supers,
             isfigures,
         )
